Created full installation script (still needs to install Kokkos & configure ReliQ). The primary workhorse of the installer is a Python script "install" located under "build". When "install" is executed under "build", a "deps" directory is created that installs Nim, UPC++, and Kokkos.
On the source side, I managed to successfully get the partitioning of the Grid into tasks working with UPC++. There's still quite a bit to add to the lattice types. One of the particularly large tasks that needs to be completed is to figure out how to have the lattice define a factory function that constructs new field objects and assigns their data according to the indices on a particular rank.

I have also started through how to partition each shared memory Grid using the functionalities provided Kokkos. This is much more subtle, but Kokkos so far has been very user-friendly, especially compared to SYCL.