This is a pretty huge update.
First, I've added "forevery" to kokkosthreads ---> kokkosdispatch, which iterates through a code blocks executing on unique threads through a vector loop. I may be incorrect, but I believe that said vector loop with vectorize on SIMD instructions on a CPU or warps/wavefronts on a GPU. In other words, each thread executes on single instructions. I'd like to change this in the future and have the SIMD instead be part of the layout, as that's the only way that we can guarantee 100% SIMD efficiency. You can see from earlier prototype versions of ReliQ that we started out with this; however, I had a hard time getting Kokkos views to mesh with said data types. I think that it is possible with some tweaking, as the Kokkos examples show that it is possible, but one needs to figure out how to get the views to operator on <SIMDNamespace>::storage_type, and I've failed at getting it to work within Nim so far. I think we'll figure it out eventually.

Second, and much more interesting, is I have succeeded in introducing Chapel-like promotion/fusion of arithematic expressions. If one writes something like "fieldA := <arbitraryArithematicExpressionOfFieldsAndScalars>", then the "promote" macro under field/simplecubicfield takes the syntax tree generated by the latter expression and transforms it into a syntax tree containing 1.) declarations of downcasted views of the involved fields and b.) morphs the AST of "... := ..." into "forevery ...: ... = ..." with the appropriate accessors on the views. I think that said macros/templates under field/simplecubicfield have been documented well enough, so as to give you a better idea of what I mean. To brag a little bit, this is honestly the coolest macro I've ever written. Practically speaking, promotion both completely eliminates the need to create temporary storage for each binary operation and pass through the data with a parallel loop for each binary operation. Instead, only one parallel pass is performed and ***no*** temporaries for the entire field are created.